+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/ForToStatementNode.java
@@ -8,6 +8,19 @@ public final class ForToStatementNode extends ForStatementNode {
     super(token);
   }
 
+  public ExpressionNode getInitializerExpression() {
+    return (ExpressionNode) jjtGetChild(2);
+  }
+
+  public ExpressionNode getTargetExpression() {
+    return (ExpressionNode) jjtGetChild(4);
+  }
+
+  @Override
+  public StatementNode getStatement() {
+    return (StatementNode) jjtGetChild(6);
+  }
+
   @Override
   public <T> T accept(DelphiParserVisitor<T> visitor, T data) {
     return visitor.visit(this, data);