+++ src/main/java/org/sonar/plugins/delphi/symbol/declaration/VariableNameDeclaration.java
@@ -1,35 +1,52 @@
 package org.sonar.plugins.delphi.symbol.declaration;
 
+import java.util.List;
 import java.util.Objects;
+import net.sourceforge.pmd.lang.ast.Node;
 import org.jetbrains.annotations.NotNull;
+import org.sonar.plugins.delphi.antlr.ast.node.ConstStatementNode;
+import org.sonar.plugins.delphi.antlr.ast.node.ForInStatementNode;
+import org.sonar.plugins.delphi.antlr.ast.node.ForLoopVarDeclarationNode;
+import org.sonar.plugins.delphi.antlr.ast.node.ForToStatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.NameDeclarationNode;
+import org.sonar.plugins.delphi.antlr.ast.node.TypeNode;
+import org.sonar.plugins.delphi.antlr.ast.node.VarStatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.Visibility;
 import org.sonar.plugins.delphi.symbol.SymbolicNode;
+import org.sonar.plugins.delphi.symbol.resolve.TypeInferrer;
 import org.sonar.plugins.delphi.symbol.scope.DelphiScope;
+import org.sonar.plugins.delphi.type.DelphiType;
 import org.sonar.plugins.delphi.type.Type;
+import org.sonar.plugins.delphi.type.Type.ArrayConstructorType;
 import org.sonar.plugins.delphi.type.Typed;
+import org.sonar.plugins.delphi.type.factory.TypeFactory;
 import org.sonar.plugins.delphi.type.generic.TypeSpecializationContext;
+import org.sonar.plugins.delphi.type.intrinsic.IntrinsicType;
 
 public final class VariableNameDeclaration extends AbstractDelphiNameDeclaration
     implements TypedDeclaration, Visibility {
   private final Type type;
   private final VisibilityType visibility;
+  private final boolean inline;
   private int hashCode;
 
   public VariableNameDeclaration(NameDeclarationNode node) {
-    this(new SymbolicNode(node), extractType(node), extractVisibility(node));
+    this(new SymbolicNode(node), extractType(node), extractVisibility(node), extractInline(node));
   }
 
   private VariableNameDeclaration(String image, Type type, DelphiScope scope) {
     super(SymbolicNode.imaginary(image, scope));
     this.type = type;
     this.visibility = VisibilityType.PUBLIC;
+    this.inline = false;
   }
 
-  private VariableNameDeclaration(SymbolicNode location, Type type, VisibilityType visibility) {
+  private VariableNameDeclaration(
+      SymbolicNode location, Type type, VisibilityType visibility, boolean inline) {
     super(location);
     this.type = type;
     this.visibility = visibility;
+    this.inline = inline;
   }
 
   public static VariableNameDeclaration compilerVariable(
@@ -38,24 +55,76 @@ public final class VariableNameDeclaration extends AbstractDelphiNameDeclaration
   }
 
   private static Type extractType(NameDeclarationNode node) {
-    Typed typed;
-
     switch (node.getKind()) {
       case CONST:
+        return constType(node);
       case EXCEPT_ITEM:
       case RECORD_VARIANT_TAG:
-        typed = (Typed) node.jjtGetParent();
-        break;
+        return ((Typed) node.jjtGetParent()).getType();
       case PARAMETER:
       case FIELD:
       case VAR:
-        typed = (Typed) node.getNthParent(2);
-        break;
+        return ((Typed) node.getNthParent(2)).getType();
+      case INLINE_CONST:
+        return inlineConstType(node);
+      case INLINE_VAR:
+        return inlineVarType(node);
+      case LOOP_VAR:
+        return loopVarType(node);
       default:
         throw new AssertionError("Unhandled DeclarationKind");
     }
+  }
 
-    return typed.getType();
+  private static Type constType(NameDeclarationNode node) {
+    Type type = ((Typed) node.jjtGetParent()).getType();
+    if (type.isArrayConstructor()) {
+      List<Type> elementTypes = ((ArrayConstructorType) type).elementTypes();
+      Type elementType = elementTypes.stream().findFirst().orElse(DelphiType.voidType());
+      TypeFactory typeFactory = node.getTypeFactory();
+      if (elementType.isInteger()) {
+        elementType = typeFactory.getIntrinsic(IntrinsicType.BYTE);
+      }
+      type = typeFactory.set(elementType);
+    }
+    return type;
+  }
+
+  private static Type inlineConstType(NameDeclarationNode node) {
+    var constStatement = (ConstStatementNode) node.jjtGetParent();
+    return getDeclaredTypeWithTypeInferenceFallback(
+        node.getTypeFactory(), constStatement.getTypeNode(), constStatement.getExpression());
+  }
+
+  private static Type inlineVarType(NameDeclarationNode node) {
+    var varStatement = (VarStatementNode) node.getNthParent(2);
+    return getDeclaredTypeWithTypeInferenceFallback(
+        node.getTypeFactory(), varStatement.getTypeNode(), varStatement.getExpression());
+  }
+
+  private static Type loopVarType(NameDeclarationNode node) {
+    ForLoopVarDeclarationNode loopVarDeclaration = (ForLoopVarDeclarationNode) node.jjtGetParent();
+    Node loop = loopVarDeclaration.jjtGetParent();
+
+    Typed typed;
+    if (loop instanceof ForToStatementNode) {
+      typed = ((ForToStatementNode) loop).getInitializerExpression();
+    } else {
+      typed = ((ForInStatementNode) loop).getCurrentDeclaration();
+    }
+
+    return getDeclaredTypeWithTypeInferenceFallback(
+        node.getTypeFactory(), loopVarDeclaration.getTypeNode(), typed);
+  }
+
+  private static Type getDeclaredTypeWithTypeInferenceFallback(
+      TypeFactory typeFactory, TypeNode typeNode, Typed typed) {
+    if (typeNode != null) {
+      return typeNode.getType();
+    } else {
+      TypeInferrer inferrer = new TypeInferrer(typeFactory);
+      return inferrer.infer(typed);
+    }
   }
 
   private static VisibilityType extractVisibility(NameDeclarationNode node) {
@@ -75,6 +144,17 @@ public final class VariableNameDeclaration extends AbstractDelphiNameDeclaration
     return visibility.getVisibility();
   }
 
+  private static boolean extractInline(NameDeclarationNode node) {
+    switch (node.getKind()) {
+      case INLINE_CONST:
+      case INLINE_VAR:
+      case LOOP_VAR:
+        return true;
+      default:
+        return false;
+    }
+  }
+
   @Override
   @NotNull
   public Type getType() {
@@ -86,9 +166,13 @@ public final class VariableNameDeclaration extends AbstractDelphiNameDeclaration
     return visibility;
   }
 
+  public boolean isInline() {
+    return inline;
+  }
+
   @Override
   protected DelphiNameDeclaration doSpecialization(TypeSpecializationContext context) {
-    return new VariableNameDeclaration(getNode(), type.specialize(context), visibility);
+    return new VariableNameDeclaration(getNode(), type.specialize(context), visibility, inline);
   }
 
   @Override